Bring in a JLine-based console in a modular way
Replace the old code that tried to use reflection to load a JLine
console with a new "rhino-cli" module that includes JLine. If this
module is present and we are on an interactive terminal, we will use a
JLine console which supports command history and editing.
Include this module in rhino-all so that the default Rhino CLI uses it.