Avoid the need for activation frames when using the debugger
The requirement to use activation frames when the debugger is in operation can cause a significant slowdown, and has previously been a way in which behavioral differences can creep in. This change refactors things to use a proxy scriptable object on the assumption that debug operations are less common than interpreter operations.