Simpler JLine integration
Move the JLine integration back into the "rhino-tools" module, but make
it optional using "compileOnly". The module will use JLine if it's in
the classpath and will fall back gracefully if it's not.

Include JLine in the "rhino-all" module so that the "rhino-all" JAR
contains JLine by default so that we have a nice initial user
experience.

This will give us command-line editing and history. Command completion
and highlighting are a task for another day.